pyproject.toml
[project]
name = "cursor-voice-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.2",
    "pyttsx3>=2.99",
    "speechrecognition>=3.15.2",
]
